Modified open and create
Changed the open to look up hash path using the dictionary. If being
opened with read only, no need to create copy of file as it will not be
modified. If write ability, a copy of the file is created, and the copy
will be opened.

For create, file is made with a temp file name of 0_random, then opened.