Fixed Spacing issues, modified open and create
Sorry for the learning of Github for previous commit and revert.

Modified open to open the file after looking up its hash path using the
dict. If the file is opened using read only, no need to create copy
prior to as it will not be modified. If it will be opened with write
abilities, copy is made by concatenating random value to path and a copy
of file made, then opens the copy.

For create, file created will be assigned placeholder file name of
0_random with random being a random value.